CatchUp

+++ Dockerfiles/agent/entrypoint/51-docker.sh
@@ -5,7 +5,7 @@
 #     (in that case, we trust the user wants docker integration and don't check existence)
 #   - we find the docker socket at it's default location
 
-if [[ -z "${DOCKER_HOST}" && ! -e /var/run/docker.sock ]]; then
+if [[ -z "${DOCKER_HOST}" && ! -S /var/run/docker.sock ]]; then
     exit 0
 fi
 
@@ -19,7 +19,7 @@ if [[ ! -e /etc/datadog-agent/datadog.yaml ]]; then
 fi
 
 # Enable the docker corecheck
-if [[ ! -e /etc/datadog-agent/conf.d/docker.d/conf.yaml.default ]]; then
+if [[ ! -e /etc/datadog-agent/conf.d/docker.d/conf.yaml.default && -e /etc/datadog-agent/conf.d/docker.d/conf.yaml.example ]]; then
     mv /etc/datadog-agent/conf.d/docker.d/conf.yaml.example \
     /etc/datadog-agent/conf.d/docker.d/conf.yaml.default
 fi
